World Edits

+++ Vanilla/Vanilla.World.Old/Network/WorldRouter.cs
@@ -9,7 +9,7 @@ namespace Vanilla.World.Network
     {
         public override WorldOpcodes FetchOpcode(PacketReader packet)
         {
-            return (WorldOpcodes)packet.ReadByte();
+            return (WorldOpcodes)packet.ReadUInt16();
         }
     }
 }
